requirements.txt
google-genai
openai
fal-client
requests
python-dotenv
cloudinary
Pillow
yt-dlp
